Move Script, PyErrorWidget and PyCellViewer into isobar to match namespace and resolve dependency order
