+++ Seabreeze/src/PySeabreezeEngine.cpp
@@ -16,7 +16,7 @@
 
 #include <functional>
 #include "hurricane/isobar/PyCell.h"
-#include "hurricane/viewer/PyCellViewer.h"
+#include "hurricane/isobar/PyCellViewer.h"
 #include "hurricane/isobar/PyNet.h"
 #include "hurricane/viewer/ExceptionWidget.h"
 #include "hurricane/Cell.h"
