+++ foehn/src/PyFoehnEngine.cpp
@@ -18,7 +18,7 @@
 #include "hurricane/isobar/PyNet.h"
 #include "hurricane/isobar/PyCell.h"
 #include "hurricane/isobar/PyInstance.h"
-#include "hurricane/viewer/PyCellViewer.h"
+#include "hurricane/isobar/PyCellViewer.h"
 #include "hurricane/viewer/ExceptionWidget.h"
 #include "hurricane/Cell.h"
 #include "crlcore/Utilities.h"
