+++ hurricane/src/configuration/meson.build
@@ -13,7 +13,7 @@ mocs = qt.compile_moc(
 resources = qt.compile_resources(
   sources: 'Configuration.qrc'
 )
-configuration = library(
+configuration = shared_library(
   'configuration',
   'Parameter.cpp',
   'LayoutDescription.cpp',
@@ -27,11 +27,11 @@ configuration = library(
   mocs,
   resources,
   dependencies: [qt_deps, boost],
-  link_with: [hurricane, isobar],
+  link_with: [hurricane, isobar, utils],
   include_directories: includes
 )
 
-pytypemanager = library(
+pytypemanager = shared_library(
   'pytypemanager',
   'ProxyProperty.cpp',
   'PyTypeManager.cpp',
@@ -40,3 +40,34 @@ pytypemanager = library(
   include_directories: includes
 )
 
+py.extension_module(
+  'Cfg',
+  'PyParameter.cpp',
+  'PyConfiguration.cpp',
+  'PyLayoutDescription.cpp',
+  'PyCfg.cpp',
+  dependencies: [qt_deps, py_deps, boost],
+  link_with: [isobar, hurricane, configuration, pytypemanager],
+  include_directories: includes
+) 
+
+py.extension_module(
+  'Hurricane3',
+  'PyDbU.cpp',
+  'PyPoint.cpp',
+  'PyBox.cpp',
+  'PyDataBase.cpp',
+  'PyTechnology.cpp',
+  'PyLayerMask.cpp',
+  'PyLayer.cpp',
+  'PyMaterial.cpp',
+  'PyBasicLayer.cpp',
+  'PyViaLayer.cpp',
+  'PyRegularLayer.cpp',
+  'PyDiffusionLayer.cpp',
+  'PyTransistorLayer.cpp',
+  'PyHurricane3.cpp',
+  dependencies: hurricane_deps,
+  link_with: [hurricane, configuration, pytypemanager],
+  include_directories: includes
+)
