+++ hurricane/src/isobar/hurricane/isobar/PyCellViewer.h
@@ -53,4 +53,4 @@ extern "C" {
 }  // End of extern "C".
 
 
-}  // End of Isobar namespace.
+}  // End of Hurricane namespace.
