+++ tramontana/src/TramontanaEngine.cpp
@@ -37,7 +37,7 @@
 #include "hurricane/Vertical.h"
 #include "hurricane/Horizontal.h"
 #include "hurricane/RoutingPad.h"
-#include "hurricane/viewer/Script.h"
+#include "hurricane/isobar/Script.h"
 #include "crlcore/Measures.h"
 #include "crlcore/Utilities.h"
 #include "crlcore/AllianceFramework.h"
