+++ unicorn/src/UnicornGui.cpp
@@ -20,9 +20,9 @@
 #include "hurricane/DebugSession.h"
 #include "hurricane/Warning.h"
 #include "hurricane/DataBase.h"
-#include "hurricane/viewer/Script.h"
+#include "hurricane/isobar/Script.h"
 #include "hurricane/viewer/CellViewer.h"
-#include "hurricane/viewer/PyCellViewer.h"
+#include "hurricane/isobar/PyCellViewer.h"
 #include "crlcore/Utilities.h"
 #include "crlcore/Catalog.h"
 #include "crlcore/AllianceFramework.h"
